% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{checkAnalysisType}
\alias{checkAnalysisType}
\title{Check that the analysis_type is valid}
\usage{
checkAnalysisType(analysis_type)
}
\arguments{
\item{analysis_type}{A string specifying the type of analysis to be done}
}
\description{
Check that the analysis_type is valid
}

