% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{computeDF}
\alias{computeDF}
\title{Function for computing the residual degrees of freedom.}
\usage{
computeDF(RCB.asr)
}
\arguments{
\item{RCB.asr}{A fitted ASReml object}
}
\description{
Function for computing the residual degrees of freedom.
}

