% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCB_SupportFunctions.R
\name{deltaAnalysis}
\alias{deltaAnalysis}
\title{Computes the deltas between the treatment means and associated p-values}
\usage{
deltaAnalysis(RCB_asr, alpha, total_df)
}
\arguments{
\item{RCB_asr}{A fitted ASReml object}

\item{alpha}{A number between 0 and 1 specifying the confidence level}

\item{total_df}{A number specifying the total degrees of freedom to be used in the p-value computations.}
}
\value{
A dataframe with columns for head, check, difference between the mean, associated p-value for all combinations of heads and checks, degree of freedom for t-test, test of statistic, lower and upper confidence intervals
}
\description{
Computes the deltas between the treatment means and associated p-values
}
