% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunction.R
\name{R.ASReml_RCB_Return}
\alias{R.ASReml_RCB_Return}
\title{Main function for doing RCBD analysis.}
\usage{
R.ASReml_RCB_Return(data, analysis_type, alpha = 0.05,
  data_fields = list(CROP_OBSRVTN_DETAIL_ID = "NUMVALUE", FIELD_ID =
  "FIELDID", FACTOR_1 = "FACTOR1", REP_ID = "REPID", Factortype_name =
  "EXPERIMENTALUNITID"))
}
\arguments{
\item{data}{A dataframe with trial observation data. Column names should match those specified as the values in \code{data_fields}.}

\item{analysis_type}{A string specifying the type of analysis to be conducted.}

\item{alpha}{A number between 0 and 1 specifying the confidence level}

\item{data_fields}{A list specifying the environmental variables to be used in fitting the model and the associated column names in the data dataframe.}
}
\value{
A list with a component for the fit model, the LS means analysis, and the delta means analysis
}
\description{
Main function for doing RCBD analysis.
}

