% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{reformatData}
\alias{reformatData}
\title{Reformats data to play nice with ASReml}
\usage{
reformatData(data, data_fields)
}
\arguments{
\item{data}{A dataframe}

\item{data_fields}{A list of column names}
}
\description{
Reformats data to play nice with ASReml
}

