% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCB_SupportFunctions.R
\name{setFixedRandomEffects}
\alias{setFixedRandomEffects}
\title{Sets the fixed and random formulas based on the analysis type}
\usage{
setFixedRandomEffects(analysis_type)
}
\arguments{
\item{analysis_type}{A string specifying the type of analysis to be done}
}
\description{
Sets the fixed and random formulas based on the analysis type
}
