% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{setLicense}
\alias{setLicense}
\title{Sets the location of the ASReml license file if it does not exist}
\usage{
setLicense()
}
\description{
Sets the location of the ASReml license file if it does not exist
}

