% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCB_SupportFunctions.R
\name{checkData}
\alias{checkData}
\title{Checks that data conditions are met by making sure the data frame has the correctly named columns}
\usage{
checkData(data, data_fields)
}
\arguments{
\item{data}{A dataframe}

\item{data_fields}{A list of column names}
}
\description{
Checks that data conditions are met by making sure the data frame has the correctly named columns
}
