% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{deltaAnalysis}
\alias{deltaAnalysis}
\title{Computes the deltas between the treatment means and associated p-values}
\usage{
deltaAnalysis(RCB.asr, total_df = 150.8)
}
\arguments{
\item{RCB.asr}{A fitted ASReml object}

\item{total_df}{A number specifying the total degrees of freedom to be used in the p-value computations.}
}
\value{
A dataframe with columns for head, check, difference between the mean, and associated p-value for all combinations of heads and checks.
}
\description{
Computes the deltas between the treatment means and associated p-values
}

