% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lsmAnalysis}
\alias{lsmAnalysis}
\title{Computes the LS means analysis}
\usage{
lsmAnalysis(RCB.asr, data, alpha)
}
\arguments{
\item{RCB.asr}{A fitted ASReml object}

\item{data}{A dataframe that contains the data used to create the ASReml object}

\item{alpha}{A number between 0 and 1 specifying the confidence level}
}
\value{
A dataframe containing the Entry, Yield, SE, and upper and lower confidence intervals.
}
\description{
Computes the LS means analysis
}

